Moved AsteroidsGame class to its own file
